+++ SurfaceTopography/Container/IO/__init__.py
@@ -22,6 +22,12 @@
 # SOFTWARE.
 #
 
+import io
+import os
+import requests
+
+from ...Exceptions import CannotDetectFileFormat, UnknownFileFormat
+
 from .CE import CEReader, write_containers
 
 readers = [
@@ -33,22 +39,18 @@ for reader in readers:
     lookup_reader_by_format[reader.format()] = reader
 
 
-def detect_format(fobj, comm=None):
+def detect_format(fobj):
     """
     Detect file format based on its content.
 
     Arguments
     ---------
     fobj : filename or file object
-    comm : mpi communicator, optional
     """
     msg = ""
     for reader in readers:
         try:
-            if comm is not None:
-                reader(fobj, comm)
-            else:
-                reader(fobj)
+            reader(fobj)
             return reader.format()
         except Exception as err:
             msg += "tried {}: \n {}\n\n".format(reader.__name__, err)
@@ -60,82 +62,23 @@ def detect_format(fobj, comm=None):
     raise CannotDetectFileFormat(msg)
 
 
-def open_topography(fobj, format=None, communicator=None):
+def open_container(fobj, format=None):
     r"""
-    Returns a reader object for the file `fobj`. The reader interface mirrors
-    the topography interface and can be used to extract meta data (number of
-    grid points, physical sizes, etc.) without reading the full topography in
-    memory.
+    Returns a container reader object for the file `fobj`.
 
     Parameters
     ----------
     fobj : str or filelike object
-        path of the file or filelike object
+        Path of the file or file-like object.
     format : str, optional
-        specify in which format the file should be interpreted
-    communicator : mpi4py or NuMPI communicator object
-        MPI communicator handling inter-process communication
+        Specify in which format the file should be interpreted.
+        (Default: None, which means autodetect file format)
 
     Returns
     -------
-    Instance of a :class:`SurfaceTopography.IO.Reader.ReaderBase` subclass
-    according to the format.
-
-    Examples
-    --------
-    Simplest read workflow:
-
-    >>> reader = open_topography("filename")
-    >>> top = reader.topography()
-
-    The first topography in file is returned, independently of whether
-    the file has multiple channels or not.
-
-    You can always check the channels and their metadata with
-    `reader.channels()`. This returns a list of dicts containing attributes `name`,
-    `physical_sizes`, `nb_grid_pts`, ``unit` and `height_scale_factor:
-
-    >>> reader.channels
-    [{'name': 'ZSensor',
-      'nb_grid_pts': (256, 256),
-      'physical_sizes': (9999.999999999998, 9999.999999999998),
-      'unit': 'nm',
-      'height_scale_factor': 0.29638271279074097},
-     {'name': 'AmplitudeError',
-      'nb_grid_pts': (256, 256),
-      'physical_sizes': (10.0, 10.0),
-      'unit': ('µm', None),
-      'height_scale_factor': 0.04577566528320313}]
-
-    Here the channel 'ZSensor' offers a topography with sizes of
-    10000 nm in each dimension.
-
-    You can choose it by giving the index 0 to channel
-    (you would use `channel=1` for the second):
-
-    >>> top = reader.topography(channel=0)
-
-    The returned topography has the physical sizes found in the file.
-
-    You can also prescribe some attributes when reading the topography:
-
-    >>> top = reader.topography(channel=0, physical_sizes=(10.,10.), info={"unit":"µm"})
-
-    In order to plot the topography with matplotlib, you can use
-
-    >>> plt.pcolormesh(*top.positions_and_heights())
-
-    with origin in the lower left and correct tick labels at x and y axes, or
-
-    >>> plt.imshow(top.heights().T)
-
-    with origin in the upper left (inverted y axis).
+    Instance of a :class:`SurfaceTopography.Container.IO.Reader.ContainerReaderBase`
+    subclass according to the format.
     """  # noqa: E501
-    if communicator is not None:
-        kwargs = {"communicator": communicator}
-    else:
-        kwargs = {}
-
     if not hasattr(fobj, 'read'):  # fobj is a path
         if not os.path.isfile(fobj):
             raise FileExistsError("file {} not found".format(fobj))
@@ -144,7 +87,7 @@ def open_topography(fobj, format=None, communicator=None):
         msg = ""
         for reader in readers:
             try:
-                return reader(fobj, **kwargs)
+                return reader(fobj)
             except Exception as err:
                 msg += "tried {}: \n {}\n\n".format(reader.__name__, err)
             finally:
@@ -156,63 +99,30 @@ def open_topography(fobj, format=None, communicator=None):
     else:
         if format not in lookup_reader_by_format.keys():
             raise UnknownFileFormat(
-                "{} not in registered file formats {}".format(
+                "{} not in registered container file formats {}".format(
                     fobj, lookup_reader_by_format.keys()))
-        return lookup_reader_by_format[format](fobj, **kwargs)
+        return lookup_reader_by_format[format](fobj)
 
 
-def read_topography(fn, format=None, communicator=None, **kwargs):
+def read_container(fn, format=None, **kwargs):
     r"""
-    Returns a topography object representing the topograpgy in the file `fobj`.
-    If there are multiple data channels within this file, the default channel
-    is returned. The default channel depends on the file format; see
-    documentation of the respective reader on this.
+    Returns a surface container object representing a list of topographies.
 
     Parameters
     ----------
     fobj : str or filelike object
-        path of the file or filelike object
-    format : str, optional
-        specify in which format the file should be interpreted
-    communicator : mpi4py or NuMPI communicator object
-        MPI communicator handling inter-process communication
-    channel : int
-        Number of the channel to load. See also `channels` method.
-    physical_sizes : tuple of floats
-        Physical size of the topography. It is necessary to specify this
-        if no physical size is found in the data file. If there is a
-        physical size in the file, then specifying this parameter will raise
-        an exception.
-    height_scale_factor : float
-        Can be used to set height scale factor if not found in the data file.
-    info : dict
-        This dictionary will be appended to the info dictionary returned
-        by the reader.
-    periodic: bool
-        Wether the SurfaceTopography should be interpreted as one period of a
-        periodic surface. This will affect the PSD and autocorrelation
-        calculations (windowing)
-    subdomain_locations : tuple of ints
-        Origin (location) of the subdomain handled by the present MPI process.
-    nb_subdomain_grid_pts : tuple of ints
-        Number of grid points within the subdomain handled by the present
-        MPI process.
+        Path of the file or file-like object.
 
     Returns
     -------
-    topography : subclass of :obj:`HeightContainer`
-        The object containing the actual topography data.
-
-    Raises
-    ------
-    MetadataAlreadyDefined
-        Raised if given arguments for `physical_sizes` or `height_scale_factor`
-        although it's already given in the file.
+    container : list of subclasses of :obj:`SurfaceContainer`
+        A list of container objects.
     """
-    with open_topography(fn, format=format,
-                         communicator=communicator) as reader:
-        t = reader.topography(**kwargs)
-    return t
+    containers = []
+    with open_container(fn, format=format) as reader:
+        for i in range(reader.nb_surfaces):
+            containers += [reader.container(index=i, **kwargs)]
+    return containers
 
 
 def read_published_container(publication_url, **request_args):